+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed small grammar mistakes in recognizing some statements
 - Updated semantic token generation logic
 - Updated logic to get file paths
+- Configuration options to let user have empty setting for patterns
+- Formatting logic for patterns to only format standalone pattern.
 
 ## [0.2.2] - 2025-03-06